Put in place the class structure
The Parse class parse the files and store the result in a list of File
class objects.
The File class store a list of Entry class objects (num of line + ID) for
the entries which don't have a translation, and a second for entries which
don't have a text in the source language.