Fix bug with parsing when destination=NONE
The minimum is also -1 so the while loop condition i < min is true but
we are not at the good column (which doesn't exist).
So I have had the condition i == min to analyze the expression.
But if if i > min, the expression is not analyzed, so we must not do i++.