Add fake translation and loanwords
A translation is missing not only when there is no text.
We can have fake translation (listed in a file), it is not the translation
of the entry, for example FRENCH.
We can have the copy of the source language. But there is loanwords (words
in different languages like OK), so there are listed in a file in order
to be not detected.